requirements.txt
anthropic
geopandas
shapely
pandas
pyproj
python-docx
openpyxl
